adding exist check

+++ etl_funcs/file_helpers.py
@@ -6,7 +6,8 @@ def cleanup_the_house(dir_lst):
     """Given a list of directories, flush dir and subdirs."""
 
     for dir_to_remove in dir_lst:
-        shutil.rmtree(dir_to_remove)
+        if os.path.exists(dir_to_remove):
+            shutil.rmtree(dir_to_remove)
 
     return
 
@@ -15,6 +16,7 @@ def make_dirs(dir_lst):
     """Given a list of directories, make empty dirs."""
 
     for dir_to_make in dir_lst:
-        os.mkdir(dir_to_make)
+        if not os.path.exists(dir_to_make):
+            os.mkdir(dir_to_make)
 
     return
